+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 0.20.49
 
+* `Repos`:
+    * `Common`:
+        * All `Repo`s get `diff` extensions
+        * `KeyValueRepo` and `KeyValuesRepo` get `applyDiff` extension
+
 ## 0.20.48
 
 * `Versions`: