+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     * `Common`:
         * All `Repo`s get `diff` extensions
         * `KeyValueRepo` and `KeyValuesRepo` get `applyDiff` extension
+        * Add new extensions `on*` flows for crud repos
 
 ## 0.20.48